requirements.txt
future
emoji
requests
sqlalchemy
python-telegram-bot==11.1.0
psycopg2-binary
pyowm
feedparser
geopy
speedtest-cli>=2.0.2
bs4
lxml
hurry.filesize

